Changed simple rules and selectors to be predicates
