+++ lib/arch_test.dart
@@ -4,9 +4,9 @@ import 'dart:io';
 
 import 'src/core/core.dart';
 
-export 'src/rules/rules.dart';
 export 'src/core/core.dart'
     hide PackageLoader, ArchTestsRunner, ArchTestDeclarator;
+export 'src/predicates/predicates.dart';
 export 'src/selectors/selectors.dart';
 
 void archTest(ArchTest test) {
