--- lib/src/rules/rules.dart
@@ -1,1 +1,0 @@
-export 'element_rules.dart';
